+++ chainermn/functions/__init__.py
@@ -1,2 +1,2 @@
-from chainermn.functions.communication import Recv  # NOQA
-from chainermn.functions.communication import Send  # NOQA
+from chainermn.functions.point_to_point_communication import Recv  # NOQA
+from chainermn.functions.point_to_point_communication import Send  # NOQA
