+++ chainermn/functions/point_to_point_communication.py
